# docs/swagger.yaml
basePath: /v1
definitions:
  docs.CacheStats:
    properties:
      hits:
        example: 120
        type: integer
      keys:
        example: 15
        type: integer
      memoryUsed:
        example: 1.19M
        type: string
      misses:
        example: 30
        type: integer
    type: object
  docs.ClearCacheResponse:
    properties:
      deleted:
        example: 42
        type: integer
    type: object
  docs.ErrorResponse:
    properties:
      error:
        example: unauthorized
        type: string
    type: object
  docs.HealthResponse:
    properties:
      api:
        example: ok
        type: string
      cacheStats:
        $ref: '#/definitions/docs.CacheStats'
      languagetool:
        example: ok
        type: string
      redis:
        example: ok
        type: string
      websocket:
        $ref: '#/definitions/docs.WebSocketStats'
    type: object
  docs.LanguageItem:
    properties:
      code:
        example: en-US
        type: string
      longCode:
        example: en-US
        type: string
      name:
        example: English (US)
        type: string
    type: object
  docs.WebSocketStats:
    properties:
      active:
        example: 3
        type: integer
      total:
        example: 47
        type: integer
    type: object
  internal_languagetool.Category:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  internal_languagetool.CheckRequest:
    properties:
      disabledCategories:
        description: Comma-separated category IDs to suppress
        type: string
      disabledRules:
        description: Comma-separated rule IDs to suppress
        type: string
      enabledCategories:
        description: Comma-separated category IDs to enable (e.g. "STYLE,PUNCTUATION")
        type: string
      enabledOnly:
        description: When true, only the rules/categories in EnabledRules/EnabledCategories
          run
        type: boolean
      enabledRules:
        description: Comma-separated rule IDs to enable on top of the active profile
        type: string
      language:
        type: string
      level:
        description: '"default" | "picky" — picky enables extra style/punctuation
          rules'
        type: string
      text:
        type: string
    type: object
  internal_languagetool.CheckResponse:
    properties:
      cacheExpiresIn:
        type: integer
      cached:
        type: boolean
      checkedAt:
        type: string
      language:
        $ref: '#/definitions/internal_languagetool.Language'
      matches:
        items:
          $ref: '#/definitions/internal_languagetool.Match'
        type: array
    type: object
  internal_languagetool.Context:
    properties:
      length:
        type: integer
      offset:
        type: integer
      text:
        type: string
    type: object
  internal_languagetool.Language:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
  internal_languagetool.Match:
    properties:
      context:
        $ref: '#/definitions/internal_languagetool.Context'
      length:
        type: integer
      message:
        type: string
      offset:
        type: integer
      replacements:
        items:
          $ref: '#/definitions/internal_languagetool.Replacement'
        type: array
      rule:
        $ref: '#/definitions/internal_languagetool.Rule'
    type: object
  internal_languagetool.Replacement:
    properties:
      value:
        type: string
    type: object
  internal_languagetool.Rule:
    properties:
      category:
        $ref: '#/definitions/internal_languagetool.Category'
      description:
        type: string
      id:
        type: string
      issueType:
        type: string
    type: object
host: localhost:4003
info:
  contact:
    name: Studio DevHub
    url: https://github.com/studio-devhub/proofapi/issues
  description: Production-ready grammar and spell-checking API powered by LanguageTool,
    Redis, and Go.
  license:
    name: MIT
    url: https://github.com/studio-devhub/proofapi/blob/main/LICENSE
  termsOfService: https://github.com/studio-devhub/proofapi
  title: ProofAPI
  version: "1.0"
paths:
  /cache:
    delete:
      description: Deletes all cached grammar check results from Redis. Does not affect
        the languages cache.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.ClearCacheResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Clear grammar check cache
      tags:
      - cache
  /check:
    post:
      consumes:
      - application/json
      description: Submits text to LanguageTool for analysis. Results are cached in
        Redis for 5 minutes.\n\n**Maximum suggestions:** set `level=picky` and add
        `enabledCategories=STYLE,PUNCTUATION,TYPOGRAPHY,GRAMMAR,MISC,CASING`
      parameters:
      - description: Text to check
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_languagetool.CheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_languagetool.CheckResponse'
        "400":
          description: Invalid input or text too short/long
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Missing or invalid API key
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "503":
          description: LanguageTool unavailable
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Check text for grammar and spelling errors
      tags:
      - grammar
  /health:
    get:
      description: 'Returns the status of all services: API, LanguageTool, Redis,
        and WebSocket hub.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.HealthResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/docs.HealthResponse'
      summary: Health check
      tags:
      - system
  /languages:
    get:
      description: Returns all languages supported by the LanguageTool engine. Results
        are cached for 1 hour.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/docs.LanguageItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List supported languages
      tags:
      - grammar
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API key required for all endpoints except /v1/health
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"